bux fixes and improvements
