+++ geode/etm/stacker/exceptions.py
@@ -1,0 +1,8 @@
+"""
+ETM Stacker exceptions.
+"""
+
+
+class EtmStackerException(Exception):
+    """Exception raised for ETM Stacker errors."""
+    pass
